requirements-deploy.txt
fastapi
uvicorn
pydantic

pandas
numpy

google-cloud-bigquery
google-auth

httpx

openai